requirements.txt
numpy
torch~=2.2.2
pillow
torchvision
pyav
transformers~=4.40.0
google-api-python-client
datasets~=2.19.0
accelerate
python-dotenv~=1.0.1
yt-dlp[default]
pandas~=2.2.2
openai~=1.23.2